+++ Nickel.Essentials/ExternalAPIs/IMoreDifficultiesApi.cs
@@ -3,4 +3,5 @@ namespace Nickel.Essentials;
 public interface IMoreDifficultiesApi
 {
 	bool AreAltStartersEnabled(State state, Deck deck);
+	StarterDeck? GetAltStarters(Deck deck);
 }